--- src/main/java/com/zipcodewilmington/froilansfarm/Aircraft.java
@@ -1,4 +1,0 @@
-package com.zipcodewilmington.froilansfarm;
-
-public interface Aircraft {
-}
