--- src/main/java/com/zipcodewilmington/froilansfarm/Animal.java
@@ -1,11 +1,0 @@
-package com.zipcodewilmington.froilansfarm;
-
-public abstract class Animal implements Eater, NoiseMaker {
-    public void mount(){
-
-    }
-
-    public void dismount() {
-
-    }
-}
