+++ src/main/java/com/zipcodewilmington/froilansfarm/Animal/Animal.java
@@ -1,0 +1,14 @@
+package com.zipcodewilmington.froilansfarm.Animal;
+
+import com.zipcodewilmington.froilansfarm.NoiseMaker;
+import com.zipcodewilmington.froilansfarm.food.Eater;
+
+public abstract class Animal implements Eater, NoiseMaker {
+    public void mount(){
+
+    }
+
+    public void dismount() {
+
+    }
+}
