--- src/main/java/com/zipcodewilmington/froilansfarm/Botanist.java
@@ -1,6 +1,0 @@
-package com.zipcodewilmington.froilansfarm;
-
-public interface Botanist {
-
-
-}
