--- src/main/java/com/zipcodewilmington/froilansfarm/Edible.java
@@ -1,4 +1,0 @@
-package com.zipcodewilmington.froilansfarm;
-
-public interface Edible {
-}
