--- src/main/java/com/zipcodewilmington/froilansfarm/Vehicle.java
@@ -1,5 +1,0 @@
-package com.zipcodewilmington.froilansfarm;
-
-public abstract class Vehicle implements NoiseMaker,Rideable {
-
-}
