+++ src/main/java/com/zipcodewilmington/froilansfarm/food/Eater.java
@@ -1,4 +1,4 @@
-package com.zipcodewilmington.froilansfarm;
+package com.zipcodewilmington.froilansfarm.food;
 
 public interface Eater {
