+++ src/main/java/com/zipcodewilmington/froilansfarm/food/Edible.java
@@ -1,0 +1,6 @@
+package com.zipcodewilmington.froilansfarm.food;
+
+public interface Edible {
+
+
+}
