+++ src/main/java/com/zipcodewilmington/froilansfarm/person/Botanist.java
@@ -1,0 +1,6 @@
+package com.zipcodewilmington.froilansfarm.person;
+
+public interface Botanist {
+
+
+}
