+++ src/main/java/com/zipcodewilmington/froilansfarm/vehicle/Aircraft.java
@@ -1,0 +1,7 @@
+package com.zipcodewilmington.froilansfarm.vehicle;
+
+public interface Aircraft {
+
+    public void fly();
+
+}
