+++ src/main/java/com/zipcodewilmington/froilansfarm/vehicle/Vehicle.java
@@ -1,0 +1,8 @@
+package com.zipcodewilmington.froilansfarm.vehicle;
+
+import com.zipcodewilmington.froilansfarm.NoiseMaker;
+import com.zipcodewilmington.froilansfarm.Rideable;
+
+public abstract class Vehicle implements NoiseMaker, Rideable {
+
+}
